requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
cloudinary==1.34.0
sqlparse==0.5.0
tzdata==2024.1
Pillow==10.0.0
psycopg2-binary==2.9.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
drf-spectacular-sidecar==2023.8.1
gunicorn==22.0.0
whitenoise==6.5.0
python-decouple==3.8
